Stop Travis triggering manual Netlify deploys for now
We want all of the awesome Netlify gives us _except_ the aggressive CD.

Ideally, we're after a workflow of:
1. Push
2. Travis runs tests/linter/format-checks
3. If CI passes, deploy to Netlify

... but connecting Netlify to the site repo means we aggressively get
Continuous Delivery with no way to disable it.

Disabling manual Netlify deployment in Travis for now until I can come
up with a nicer solution.

This post details a workflow where Travis triggers a Netlify publish
(rather than a deploy), which may work, but keen for a cleaner solution:
https://www.laroberto.com/publishing-in-netlify-via-travisci/